# docs/openapi.yaml
# Generated by scripts/gen-openapi.sh — do not edit by hand.
# Regenerate with: make openapi
openapi: 3.1.0
info:
  title: CSV Processor
  summary: Upload, parse and archive stock-on-hand CSV exports.
  description: 'Parses CSV files in the stock-on-hand export format (`sku`, `product_name`, `price`),
    renders the parsed lines, and archives the original file to S3-compatible object storage where a lifecycle
    policy transitions it to Glacier.


    Built for a DevOps case study — see `ASSUMPTIONS.md` in the repository.'
  version: 1.1.0
paths:
  /:
    get:
      tags:
      - ui
      summary: Upload form and processed-file list
      operationId: index__get
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
  /upload:
    post:
      tags:
      - ui
      summary: Upload a CSV via the web form
      operationId: upload_form_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_form_upload_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /files/{key}:
    get:
      tags:
      - ui
      summary: Re-read and display a stored file
      operationId: view_file_files__key__get
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          title: Key
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/files:
    get:
      tags:
      - api
      summary: List previously processed files
      operationId: api_list_api_v1_files_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessedFileList'
        '502':
          description: Object storage unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - api
      summary: Upload and process a CSV file
      operationId: api_upload_api_v1_files_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_api_upload_api_v1_files_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseResultResponse'
        '400':
          description: Empty file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: File too large.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Not a CSV file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Object storage unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/files/{key}:
    get:
      tags:
      - api
      summary: Re-parse a previously processed file
      operationId: api_get_api_v1_files__key__get
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          title: Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseResultResponse'
        '404':
          description: No such object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /healthz:
    get:
      tags:
      - ops
      summary: Liveness probe
      description: 'Liveness only — deliberately does not touch S3, so a storage outage

        restarts nothing.'
      operationId: healthz_healthz_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /readyz:
    get:
      tags:
      - ops
      summary: Readiness probe
      description: Readiness — the pod cannot usefully serve traffic without object storage.
      operationId: readyz_readyz_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadyResponse'
        '503':
          description: Object storage unreachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Body_api_upload_api_v1_files_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
          description: CSV file to process.
      type: object
      required:
      - file
      title: Body_api_upload_api_v1_files_post
    Body_upload_form_upload_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
          description: CSV file to process.
      type: object
      required:
      - file
      title: Body_upload_form_upload_post
    ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
          examples:
          - File is not a CSV.
      type: object
      required:
      - detail
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HealthResponse:
      properties:
        status:
          type: string
          title: Status
          examples:
          - ok
        environment:
          type: string
          title: Environment
          examples:
          - dev
      type: object
      required:
      - status
      - environment
      title: HealthResponse
    ParseErrorModel:
      properties:
        line_number:
          type: integer
          title: Line Number
        reason:
          type: string
          title: Reason
          examples:
          - expected 3 columns, got 2
        raw:
          type: string
          title: Raw
          description: Truncated copy of the offending line.
      type: object
      required:
      - line_number
      - reason
      - raw
      title: ParseErrorModel
      description: A line that could not be parsed. The rest of the file is still processed.
    ParseResultResponse:
      properties:
        key:
          type: string
          title: Key
        filename:
          type: string
          title: Filename
        columns:
          items:
            type: string
          type: array
          title: Columns
          examples:
          - - sku
            - product_name
            - price
        row_count:
          type: integer
          title: Row Count
          description: Total data rows in the file.
          examples:
          - 750
        displayed_rows:
          type: integer
          title: Displayed Rows
          description: Rows included in this response (capped by APP_MAX_ROWS_DISPLAY).
        truncated:
          type: boolean
          title: Truncated
          description: True when the file has more rows than were returned.
        had_header:
          type: boolean
          title: Had Header
        errors:
          items:
            $ref: '#/components/schemas/ParseErrorModel'
          type: array
          title: Errors
        rows:
          items:
            $ref: '#/components/schemas/ParsedRow'
          type: array
          title: Rows
      type: object
      required:
      - key
      - filename
      - columns
      - row_count
      - displayed_rows
      - truncated
      - had_header
      title: ParseResultResponse
      description: The result of parsing an uploaded or previously stored CSV.
    ParsedRow:
      properties:
        line_number:
          type: integer
          title: Line Number
          description: 1-indexed line in the source file.
          examples:
          - 1
        values:
          items:
            type: string
          type: array
          title: Values
          description: Field values in column order.
          examples:
          - - '211627629'
            - Purple Safi Kaftan
            - '4900.0000'
      type: object
      required:
      - line_number
      - values
      title: ParsedRow
      description: A single data row from the uploaded CSV.
    ProcessedFile:
      properties:
        key:
          type: string
          title: Key
          description: Object storage key.
          examples:
          - uploads/2026/08/07/9f2c-soh.csv
        filename:
          type: string
          title: Filename
          examples:
          - soh.csv
        size_bytes:
          type: integer
          title: Size Bytes
          examples:
          - 45720
        row_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Row Count
          description: Rows counted at upload time, from object metadata.
          examples:
          - 750
        uploaded_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Uploaded At
      type: object
      required:
      - key
      - filename
      - size_bytes
      title: ProcessedFile
      description: Summary of a file that has been parsed and archived to object storage.
    ProcessedFileList:
      properties:
        count:
          type: integer
          title: Count
        files:
          items:
            $ref: '#/components/schemas/ProcessedFile'
          type: array
          title: Files
      type: object
      required:
      - count
      - files
      title: ProcessedFileList
      description: Previously processed files, newest first.
    ReadyResponse:
      properties:
        status:
          type: string
          title: Status
          examples:
          - ready
        storage:
          type: string
          title: Storage
          description: Object storage reachability.
          examples:
          - reachable
      type: object
      required:
      - status
      - storage
      title: ReadyResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
tags:
- name: api
  description: JSON API for programmatic use.
- name: ui
  description: Server-rendered HTML interface.
- name: ops
  description: Health and readiness probes.